# Backend/docs/swagger.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Library-API",
        "version": "1.0.0",
        "description": "### This is a Swagger UI for our Library API  \n\n### Error handling\nErrors are presented using the HTTP status codes 400-599. The response include a more detailed explanation about the error in the body.\n\n```\n{\n  \"error\": \"Cleartext explanation of the error that occurred\"\n}\n```\n\n### Generic API errors\nThese error responses can be returned by any of the described endpoints.\n\n| HTTP status  |  Description |\n|---|---|\n| 400| Bad input parameters or invalid request. More information can be found in the response body.\n| 401| The consumer does not have access to the requested resource or the provided token is not valid.\n| 404| The requested resource can not be found or the user does not have access to the resource.\n| 5XX| A server error has occurred while processing the request."
    },
    "paths": {
...
        "/loans/": {
            "get": {
                "tags": [
                    "loans"
                ],
                "parameters": [
                    {
                        "examples": {
                            "key": {
                                "value": "30cf8c1d-3fb9-4149-a2eb-20e3e88b89d6"
                            }
                        },
                        "name": "key",
                        "description": "API key is required to access the API",
                        "schema": {
                            "type": "string"
                        },
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {}
                                },
                                "examples": {
                                    "Loan": {
                                        "value": [
                                            {
                                                "_id": "63284a920e4f9ec5457bd08c",
                                                "loanStart": 1669627070,
                                                "loanEnd": 1670231870,
                                                "book": "63283f9913c575a0084104ee",
                                                "user": "632849e813c575a0084104fa"
                                            },
                                            {
                                                "_id": "63284ac00e4f9ec5457bd091",
                                                "loanStart": 1670923070,
                                                "loanEnd": 1673947070,
                                                "book": "6328365f13c575a0084104ee",
                                                "user": "632849e813c575a0084104fa"
                                            }
                                        ]
                                    }
                                }
                            }
                        },
                        "description": "Request successful"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Missing API key error": {
                                        "value": {
                                            "error": "Missing API key"
                                        }
                                    }
                                }
                            }
                        },
                        "description": "API key is not provided"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Invalid API key error": {
                                        "value": {
                                            "error": "Invalid API key"
                                        }
                                    }
                                }
                            }
                        },
                        "description": "API key is invalid"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Something went wrong error": {
                                        "value": {
                                            "error": "Looks like something went wrong :("
                                        }
                                    }
                                }
                            }
                        },
                        "description": "Something went wrong during the query"
                    }
                },
                "summary": "Get a list of all current loans",
                "description": ""
            },
            "post": {
                "tags": [
                    "loans"
                ],
                "parameters": [
                    {
                        "examples": {
                            "key": {
                                "value": "30cf8c1d-3fb9-4149-a2eb-20e3e88b89d6\t\r\n"
                            }
                        },
                        "name": "key",
                        "description": "API key is required to access the API",
                        "schema": {
                            "type": "string"
                        },
                        "in": "query",
                        "required": true
                    },
                    {
                        "examples": {
                            "userId": {
                                "value": "632849e813c575a0084104fa"
                            }
                        },
                        "name": "user",
                        "description": "A id of a user",
                        "schema": {
                            "type": "string"
                        },
                        "in": "query",
                        "required": true
                    },
                    {
                        "examples": {
                            "bookId": {
                                "value": "632849e813c575a0084104fa"
                            }
                        },
                        "name": "book",
                        "description": "A id of a book",
                        "schema": {
                            "type": "string"
                        },
                        "in": "query",
                        "required": true
                    },
                    {
                        "examples": {
                            "loanStart": {
                                "value": "1669627070"
                            }
                        },
                        "name": "loanStart",
                        "description": "The start date of the loan",
                        "schema": {
                            "type": "number"
                        },
                        "in": "query",
                        "required": true
                    },
                    {
                        "examples": {
                            "loanEnd": {
                                "value": "1670231870"
                            }
                        },
                        "name": "loanEnd",
                        "description": "The end date of the loan",
                        "schema": {
                            "type": "number"
                        },
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Loan": {
                                        "value": "Loan added!"
                                    }
                                }
                            }
                        },
                        "description": "Request successful"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "One or all params are missing error": {
                                        "value": {
                                            "error": "One or all params are missing. Required params: loanStart, loanEnd, userId, bookId"
                                        }
                                    }
                                }
                            }
                        },
                        "description": "Missing parameters"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Missing API key error": {
                                        "value": {
                                            "error": "Missing API key"
                                        }
                                    }
                                }
                            }
                        },
                        "description": "API key is not provided"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Invalid API key error": {
                                        "value": {
                                            "error": "Invalid API key"
                                        }
                                    }
                                }
                            }
                        },
                        "description": "API key is invalid"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Something went wrong error": {
                                        "value": {
                                            "error": "Looks like something went wrong :("
                                        }
                                    }
                                }
                            }
                        },
                        "description": "Something went wrong during the query"
                    }
                },
                "summary": "Add a new loan"
            }
        },
...
        "/loans/user/:id/": {
            "get": {
                "tags": [
                    "loans"
                ],
                "parameters": [
                    {
                        "examples": {
                            "key": {
                                "value": "30cf8c1d-3fb9-4149-a2eb-20e3e88b89d6"
                            }
                        },
                        "name": "key",
                        "description": "API key is needed to access the API",
                        "schema": {
                            "type": "string"
                        },
                        "in": "query",
                        "required": true
                    },
                    {
                        "examples": {
                            "userId": {
                                "value": "632849e813c575a0084104fa"
                            }
                        },
                        "name": "id",
                        "description": "A id of a user",
                        "schema": {
                            "type": "string"
                        },
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {}
                                },
                                "examples": {
                                    "loans": {
                                        "value": [
                                            {
                                                "_id": "632852660faec33f2cf36f63",
                                                "loanStart": 1669627070,
                                                "loanEnd": 1670231870,
                                                "book": "63283f9913c575a0084104ee",
                                                "user": "632849e813c575a0084104fa"
                                            }
                                        ]
                                    }
                                }
                            }
                        },
                        "description": "Request successful"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "error": {
                                        "value": {
                                            "error": "Missing API key"
                                        }
                                    }
                                }
                            }
                        },
                        "description": "API key is missing"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "error": {
                                        "value": {
                                            "error": "Invalid API key"
                                        }
                                    }
                                }
                            }
                        },
                        "description": "API key is invalid"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "error": {
                                        "value": {
                                            "error": "Looks like we couldn't find what you were looking for."
                                        }
                                    }
                                }
                            }
                        },
                        "description": "User not found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "error": {
                                        "value": {
                                            "error": "Looks like something went wrong :("
                                        }
                                    }
                                }
                            }
                        },
                        "description": "Something went wrong during the query"
                    }
                },
                "summary": "Get a list of loans by 1 user"
            }
        },
...
    }
}